moved everything into data parser
